+++ app/src/main/java/com/example/have_it/AddHabitActivity.java
@@ -182,41 +182,9 @@ public class AddHabitActivity extends AppCompatActivity implements FirestoreAddD
             data.put("dateStart", startDateTimestamp);
             data.put("weekdayReg", weekdayReg);
             data.put("publicity", publicity);
+            HabitController.addHabit(data,habitListReference,title);
+            finish();
 
-            habitListReference
-                    .document(title)
-                    .get()
-                    .addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-                        @Override
-                        public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-                            if (task.isSuccessful()) {
-                                DocumentSnapshot document = task.getResult();
-                                if (document.exists()) {
-                                    Toast.makeText(getApplicationContext(),"cannot add because the habit with same title exists", Toast.LENGTH_LONG).show();
-                                } else {
-                                    habitListReference
-                                            .document(title)
-                                            .set(data)
-                                            .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                                @Override
-                                                public void onSuccess(Void aVoid) {
-                                                    // These are a method which gets executed when the task is succeeded
-                                                    Log.d("Adding Habit", "Habit data has been added successfully!");
-                                                    finish();
-                                                }
-                                            })
-                                            .addOnFailureListener(new OnFailureListener() {
-                                                @Override
-                                                public void onFailure(@NonNull Exception e) {
-                                                    // These are a method which gets executed if there’s any problem
-                                                    Log.d("Adding Habit", "Habit data could not be added!" + e.toString());
-                                                    Toast.makeText(getApplicationContext(),"Not being able to add data, please check duplication title", Toast.LENGTH_LONG).show();
-                                                }
-                                            });
-                                }
-                            }
-                        }
-                    });
 
         }
     }
