+++ app/src/main/java/com/example/have_it/UserLoginActivity.java
@@ -99,8 +99,10 @@ public class UserLoginActivity extends AppCompatActivity implements View.OnClick
      * Also check password and email validity, and notify the user error inputs.
      */
     private void userLogin() {
-        String email = editTextEmail.getText().toString().trim();
-        String password = editTextPassword.getText().toString().trim();
+//        String email = editTextEmail.getText().toString().trim();
+//        String password = editTextPassword.getText().toString().trim();
+        String email = "2833395006a@gmail.com";
+        String password = "123456";
 
 
         if(email.isEmpty()){
