+++ app/src/main/java/com/example/have_it/ViewEditEventActivity.java
@@ -155,22 +155,9 @@ public class ViewEditEventActivity extends AppCompatActivity implements Database
         delete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                // Toast.makeText( getApplicationContext(),  selected_event_date, Toast.LENGTH_SHORT).show();
-                eventListReference.document(selectedEventDate)
-                        .delete()
-                        .addOnSuccessListener(new OnSuccessListener<Void>() {
-                            @Override
-                            public void onSuccess(Void aVoid) {
-                                Log.d("Delete Event", "Habit data has been deleted successfully!");
-                                finish();
-                            }
-                        })
-                        .addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                Log.w("Delete event", "Error deleting document", e);
-                            }
-                        });
+                deleteDataInFirestore();
+
+
             }
         });
 
@@ -258,76 +245,17 @@ public class ViewEditEventActivity extends AppCompatActivity implements Database
             data.put("date", dateText.getText().toString());
             data.put("latitude", latitude);
             data.put("longitude", longitude);
+            String newDate=dateText.getText().toString();
 
-            if (dateText.getText().toString().equals(selectedEventDate)){
-                eventListReference
-                        .document(selectedEventDate)
-                        .update(data)
-                        .addOnSuccessListener(new OnSuccessListener<Void>() {
-                            @Override
-                            public void onSuccess(Void aVoid) {
-                                Log.d("Edit Habit", "Habit data has been deleted successfully!");
-                                finish();
-                            }
-                        })
-                        .addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                Log.w("Edit Habit", "Error deleting document", e);
-                            }
-                        });
+            if (newDate.equals(selectedEventDate)){
+                EventController.editEventWithSameDate(eventListReference,selectedEventDate,data);
+                finish();
+
             }
             else {
-                eventListReference
-                        .document(dateText.getText().toString())
-                        .get()
-                        .addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-                            @Override
-                            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-                                if (task.isSuccessful()) {
-                                    DocumentSnapshot document = task.getResult();
-                                    if (document.exists()) {
-                                        Toast.makeText(getApplicationContext(), "cannot edit event: another event at the same day", Toast.LENGTH_LONG).show();
-                                    } else {
-                                        eventListReference
-                                                .document(selectedEventDate)
-                                                .delete()
-                                                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                                    @Override
-                                                    public void onSuccess(Void aVoid) {
-                                                        Log.d("Delete event", "event data has been deleted successfully!");
+                EventController.editEventWithDiffDate( eventListReference,  selectedEventDate, data, newDate);
+                finish();
 
-                                                    }
-                                                })
-                                                .addOnFailureListener(new OnFailureListener() {
-                                                    @Override
-                                                    public void onFailure(@NonNull Exception e) {
-                                                        Log.w("Delete event", "Error deleting document", e);
-                                                    }
-                                                });
-                                        eventListReference
-                                                .document(dateText.getText().toString())
-                                                .set(data)
-                                                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                                    @Override
-                                                    public void onSuccess(Void aVoid) {
-                                                        // These are a method which gets executed when the task is succeeded
-                                                        Log.d("Adding event", "event data has been edited successfully!");
-                                                        finish();
-                                                    }
-                                                })
-                                                .addOnFailureListener(new OnFailureListener() {
-                                                    @Override
-                                                    public void onFailure(@NonNull Exception e) {
-                                                        // These are a method which gets executed if there’s any problem
-                                                        Log.d("Adding event", "Habit data could not be edited!" + e.toString());
-                                                        Toast.makeText(getApplicationContext(), "Not being able to edit data, please check duplication event", Toast.LENGTH_LONG).show();
-                                                    }
-                                                });
-                                    }
-                                }
-                            }
-                        });
             }
         }
     }
@@ -341,22 +269,8 @@ public class ViewEditEventActivity extends AppCompatActivity implements Database
                 .document(logged.getUID())
                 .collection("HabitList")
                 .document(selectedHabit).collection("EventList");
-        // Toast.makeText( getApplicationContext(),  selected_event_date, Toast.LENGTH_SHORT).show();
-        eventListReference.document(selectedEventDate)
-                .delete()
-                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                        Log.d("Delete Event", "Habit data has been deleted successfully!");
-                        finish();
-                    }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        Log.w("Delete event", "Error deleting document", e);
-                    }
-                });
+        EventController.deleteEvent(eventListReference,selectedEventDate);
+        finish();
     }
 
     @Override
